Kindrey, stop enabling bullies and abusers.
Guns of Icarus players xedeon, HelFyre, FadedSilhoutte, Dementio, SlyCrossfox, Genesis and Kruegerz are bullies. The abuser Inkjet has not only been misusing the position as Muse community manager to protect those bullies but is actively helping them in attacking their victims. Muse Games head Howard Tsao (aka Bubbles) is aware of this abuse going on but has chosen to silence me instead of rooting out the abusers and bullies in Muse and the community.